+++ docs/changes.rst
@@ -19,6 +19,7 @@ Unreleased.
  - Updated :meth:`Image.chop() <wand.image.BaseImage.chop>` method to support ``gravity`` keyword.
  - Updated :meth:`Image.extent() <wand.image.BaseImage.extent>` method to support ``gravity`` keyword. [:issue:`554`]
  - [TEST] Migrated CI from `travis-ci.org <https://travis-ci.org/github/emcconville/wand>`_ to `travis-ci.com <https://app.travis-ci.com/emcconville/wand>`_.
+ - [TEST] Removed unneeded SVG dependency from regression test.
 
 .. _changelog-0.6.7: